loadConfigModel added for each dialog

+++ includes/Generator.h
@@ -20,7 +20,7 @@ public:
     Generator(const Distribution & dist);
     virtual ~Generator() = 0;
     Distribution GetDistribution() {return  m_dist;}
-    int32_t GetDistSize() const {return m_distSize;}
+    size_t GetDistSize() const {return m_distSize;}
 
     virtual int32_t GenerateValue(double alpha) = 0;
 protected:
